tests/Dataplace.Imersao.Core.Tests/Domain/Orcamentos/Enums/TpRegistroEnumTest.cs
using Dataplace.Imersao.Core.Domain.Orcamentos.Enums;
using Xunit;

namespace Dataplace.Imersao.Core.Tests.Domain.Orcamentos.Enums
{
    public class TpRegistroEnumTest
    {

        [Fact]
        [Trait("TpRegistroEnum", "ToDataValue")]
        public void TpRegistroToDataValueTest()
        {

            // arrange act assert
            Assert.Equal("1", TpRegistroEnum.ProdutoFinal.ToDataValue());
            Assert.Equal("2", TpRegistroEnum.SubProduto.ToDataValue());
            Assert.Equal("3", TpRegistroEnum.MateriaPrima.ToDataValue());
            Assert.Equal("4", TpRegistroEnum.Componente.ToDataValue());
            Assert.Equal("5", TpRegistroEnum.Servico.ToDataValue());
            Assert.Equal("9", TpRegistroEnum.BemDeConsumo.ToDataValue());
            Assert.Equal(default, default(TpRegistroEnum));
        }

        [Fact]
        [Trait("TpRegistroEnum", "ToTpRegistroEnum")]
        public void TpRegistroToEnumValueTest()
        {

            // arrange act assert
            Assert.Equal(TpRegistroEnum.ProdutoFinal, "1".ToTpRegistroEnum());
            Assert.Equal(TpRegistroEnum.SubProduto, "2".ToTpRegistroEnum());
            Assert.Equal(TpRegistroEnum.MateriaPrima, "3".ToTpRegistroEnum());
            Assert.Equal(TpRegistroEnum.Componente, "4".ToTpRegistroEnum());
            Assert.Equal(TpRegistroEnum.Servico, "5".ToTpRegistroEnum());
            Assert.Equal(TpRegistroEnum.BemDeConsumo, "9".ToTpRegistroEnum());
            Assert.Equal(default(TpRegistroEnum), "x".ToTpRegistroEnum());
            Assert.Equal(default, null);
        }
    }
}
